refactor: MyProfile 컴포넌트를 features 디렉토리로 이동
